+++ src/decoder/symphonia.rs
@@ -5,7 +5,7 @@ use symphonia::{
         audio::{AudioBufferRef, SampleBuffer, SignalSpec},
         codecs::{Decoder, DecoderOptions, CODEC_TYPE_NULL},
         errors::Error,
-        formats::{FormatOptions, FormatReader, SeekedTo},
+        formats::{FormatOptions, FormatReader, SeekMode, SeekTo, SeekedTo},
         io::MediaSourceStream,
         meta::MetadataOptions,
         probe::Hint,
@@ -14,9 +14,11 @@ use symphonia::{
     default::get_probe,
 };
 
-use super::DecoderError;
-use crate::common::{ChannelCount, Sample, SampleRate};
-use crate::{source, Source};
+use super::{DecoderError, Settings};
+use crate::{
+    common::{ChannelCount, Sample, SampleRate},
+    source, Source,
+};
 
 // Decoder errors are not considered fatal.
 // The correct action is to just get a new packet and try again.
@@ -30,14 +32,12 @@ pub(crate) struct SymphoniaDecoder {
     total_duration: Option<Time>,
     buffer: SampleBuffer<Sample>,
     spec: SignalSpec,
+    seek_mode: SeekMode,
 }
 
 impl SymphoniaDecoder {
-    pub(crate) fn new(
-        mss: MediaSourceStream,
-        extension: Option<&str>,
-    ) -> Result<Self, DecoderError> {
-        match SymphoniaDecoder::init(mss, extension) {
+    pub(crate) fn new(mss: MediaSourceStream, settings: &Settings) -> Result<Self, DecoderError> {
+        match SymphoniaDecoder::init(mss, settings) {
             Err(e) => match e {
                 Error::IoError(e) => Err(DecoderError::IoError(e.to_string())),
                 Error::DecodeError(e) => Err(DecoderError::DecodeError(e)),
@@ -60,17 +60,22 @@ impl SymphoniaDecoder {
 
     fn init(
         mss: MediaSourceStream,
-        extension: Option<&str>,
+        settings: &Settings,
     ) -> symphonia::core::errors::Result<Option<SymphoniaDecoder>> {
         let mut hint = Hint::new();
-        if let Some(ext) = extension {
+        if let Some(ext) = settings.hint.as_ref() {
             hint.with_extension(ext);
         }
         let format_opts: FormatOptions = FormatOptions {
-            enable_gapless: true,
+            enable_gapless: settings.gapless,
             ..Default::default()
         };
         let metadata_opts: MetadataOptions = Default::default();
+        let seek_mode = if settings.coarse_seek {
+            SeekMode::Coarse
+        } else {
+            SeekMode::Accurate
+        };
         let mut probed = get_probe().format(&hint, mss, &format_opts, &metadata_opts)?;
 
         let stream = match probed.format.default_track() {
@@ -144,6 +149,7 @@ impl SymphoniaDecoder {
             total_duration,
             buffer,
             spec,
+            seek_mode,
         }))
     }
 
@@ -178,8 +184,6 @@ impl Source for SymphoniaDecoder {
     }
 
     fn try_seek(&mut self, pos: Duration) -> Result<(), source::SeekError> {
-        use symphonia::core::formats::{SeekMode, SeekTo};
-
         let seek_beyond_end = self
             .total_duration()
             .is_some_and(|dur| dur.saturating_sub(pos).as_millis() < 1);
@@ -197,7 +201,7 @@ impl Source for SymphoniaDecoder {
         let seek_res = self
             .format
             .seek(
-                SeekMode::Accurate,
+                self.seek_mode,
                 SeekTo::Time {
                     time,
                     track_id: None,